refactor: convert hash.js to ts
